Removed tensorflow from test_requirements (now torchvision). Fixed a bug with relu backprop with datatypes. Removed loops in Conv2D. Modified ImageClassification test to use a better model architecture

+++ DLL/DeepLearning/Layers/Activations/_ReLU.py
@@ -47,5 +47,5 @@ class ReLU(Activation):
         if dCdy.shape != self.input.shape:
             raise ValueError(f"dCdy is not the same shape as the spesified output_shape ({dCdy.shape[1:], self.output_shape}).")
 
-        dCdx = dCdy * (self.input > 0)
+        dCdx = dCdy * (self.input > 0).to(dCdy.dtype)
         return dCdx
